requirements.txt
APScheduler==3.6.1
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
entrypoints==0.3
flake8==3.7.8
Flask==1.1.1
Flask-APScheduler==1.11.0
Flask-SocketIO==4.2.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
gunicorn==19.9.0
httplib2==0.13.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.0
python-engineio==3.9.3
python-socketio==4.3.1
pytz==2019.2
redis==3.3.8
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
tzlocal==2.0.0
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.5
